requirements.txt
attrs==19.1.0
aws-cdk.aws-events==0.33.0
aws-cdk.aws-iam==0.33.0
aws-cdk.aws-kms==0.33.0
aws-cdk.aws-s3==0.33.0
aws-cdk.cdk==0.33.0
aws-cdk.cx-api==0.33.0
aws-cdk.region-info==0.33.0
cattrs==0.9.0
jsii==0.11.0
mypy-extensions==0.4.1
publication==0.0.3
python-dateutil==2.8.0
six==1.12.0
typing-extensions==3.7.2
